requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
cssselect2==0.8.0
distro==1.9.0
Django==5.2.7
fonttools==4.60.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
Markdown==3.9
openai==2.6.1
pillow==12.0.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydyf==0.11.0
pyphen==0.17.2
python-dotenv==1.2.1
sniffio==1.3.1
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
